+++ CHANGELOG.md
@@ -6,7 +6,14 @@
 - Updated the server install script to allow for a more flexible selection of the server install directory
 - Changed the travis configuration. Reinstated `php 5.5` testing and temporarily disabled testing for `hhvm` as there seems to be an issue surrounding phpunit
 - Fixed the `webtest.php` file to support newer versions of PHPunit.
-- Added a simple GRAPHSON 3.0 serializer
+- Fixed an issue where using authentication and sessions could fail to maintain the session
+- Added a basic GraphSON 3.0 serializer. Until gremlin server `3.2.x` is no longer maintained, bellow will be the correct way of setting the serializer up.
+  ```php
+  $db = ne Connection();
+  $db->message->registerSerializer('\Brightzone\GremlinDriver\Serializers\Gson3', TRUE);
+  ```
+  This will later be changed to the default. After which a serializer option will be added to `Connection`
+- Added a test suit for GraphSON 3.0 and modified travis to run the newly implemented suits
 
 3.0.2
 =====